requirements.txt
aioredis==1.2.0
amqp==2.4.0
appnope==0.1.0
Arpeggio==1.9.0
asgi-redis==1.4.3
asgiref==2.3.2
assertpy==0.14
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==18.2.0
autobahn==19.1.1
Automat==0.7.0
Babel==2.6.0
backcall==0.1.0
bandit==1.5.1
billiard==3.5.0.5
cached-property==1.5.1
certifi==2018.11.29
channels==2.1.6
channels-redis==2.3.2
chardet==3.0.4
constantly==15.1.0
daphne==2.2.3
decorator==4.3.2
dj-database-url==0.5.0
Django==2.0.1
django-cors-headers==1.2.2
django-extensions==2.1.2
django-heroku==0.3.1
django-rest-auth==0.6.0
django-storages==1.6.5
django-su==0.8.0
djangorestframework==3.7.7
drfdocs==0.0.9
flake8==3.6.0
flake8-commas==2.0.0
flower==0.9.2
gems==0.3.5
gitdb2==2.0.5
gunicorn==19.1.1
hiredis==1.0.0
httplib2==0.12.0
hyperlink==18.0.0
idna==2.5
incremental==17.5.0
iso==0.0.3
isort==4.3.4
jedi==0.13.2
jellyfish==0.5.6
joblib==0.13.1
kombu==4.2.0
mccabe==0.6.1
more-itertools==5.0.0
msgpack==0.5.6
msgpack-python==0.5.6
parso==0.3.2
pbr==5.1.2
pexpect==4.6.0
pickleshare==0.7.5
pluggy==0.8.1
prompt-toolkit==1.0.15
psycopg2==2.7.7
psycopg2-binary==2.7.4
ptyprocess==0.6.0
py==1.7.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycodestyle==2.4.0
pyflakes==2.0.0
Pygments==2.3.1
PyHamcrest==1.9.0
pytest==3.9.2
pytest-django==3.4.3
python-dateutil==2.7.5
python-dotenv==0.10.3
python-http-client==3.1.0
python-json-logger==0.1.9
pytz==2015.7
redis==2.10.6
requests==2.21.0
roman==2.0.0
rsa==4.0
six==1.12.0
smmap2==2.0.5
stevedore==1.30.0
tornado==5.1.1
tqdm==4.25.0
traitlets==4.3.2
Twisted==18.9.0
txaio==18.8.1
urllib3==1.22
user-agent==0.1.5
valid-model==0.4.0
vine==1.2.0
wcwidth==0.1.7
whitenoise==3.3.1
zope.interface==4.6.0
